ThemeResource and its subclass removed.
