+++ app/src/main/webapp/themes/basicmobile/theme.xml
@@ -105,6 +105,4 @@
         </rendition>
     </template>
     
-     <!-- resources -->
-    <resource path="images/forward_mobile.png" />
 </weblogtheme>
