+++ app/src/main/webapp/themes/gaurav/theme.xml
@@ -133,36 +133,5 @@
         </rendition>
     </template>
     
-    <resource path="css/font-awesome.min.css" />
-    <resource path="css/bootstrap.css" />
-    <resource path="css/style.css" />
-
-    <resource path="js/jquery.js" />
-    <resource path="js/bootstrap.js" />
-    <resource path="js/searchhi.js"/>
-	
-    <resource path="font/FontAwesome.otf" />	
-    <resource path="font/fontawesome-webfont.eot" />
-    <resource path="font/fontawesome-webfont.svg" />
-    <resource path="font/fontawesome-webfont.ttf" />
-    <resource path="font/fontawesome-webfont.woff" />
-
-    <resource path="icons/asf-logo.png" />
-    <resource path="icons/comment.gif" />
-    <resource path="icons/delicious.gif" />
-    <resource path="icons/digman.gif" />
-    <resource path="icons/facebook-logo.png" />
-    <resource path="icons/feed-icon-14x14.png" />
-    <resource path="icons/feed-icon-28x28.png" />
-    <resource path="icons/flickr-icon.png" />
-    <resource path="icons/icon_clock.gif" />
-    <resource path="icons/linkedin-logo.jpg" />
-    <resource path="icons/page_white.png" />
-    <resource path="icons/page_white_acrobat.png" />
-    <resource path="icons/slashdot-13x13.png" />
-    <resource path="icons/tag_blue.png" />
-    <resource path="icons/technorati-16x15.png" />
-    <resource path="icons/twitter-icon.png" />
-
 </weblogtheme>
